+++ TikTakToe/TikTakToe/GameView.swift
@@ -1,12 +1,11 @@
 //
-//  ContentView.swift
+//  GameView.swift
 //  TikTakToe
 //
 //  Created by Amarjit on 05/05/2025.
 
 
 import SwiftUI
-import Combine
 
 enum GameState: Int {
     case inProgress, humanWin, computerWin, draw
@@ -20,6 +19,31 @@ enum Player: Int, CaseIterable {
     }
 }
 
+struct AlertItem: Identifiable {
+    let id: UUID
+    let title: Text
+    let message: Text
+    let buttonTitle: Text
+}
+
+struct AlertContext {
+    static let humanWin = AlertItem(id: UUID(),
+                             title: Text("You Win!"),
+                             message: Text("You beat the computer. Well done!"),
+                             buttonTitle: Text("Play again"))
+    
+    static let computerWin = AlertItem(id: UUID(),
+                             title: Text("You Lost!"),
+                             message: Text("Better luck next time"),
+                             buttonTitle: Text("Play again"))
+    
+    static let draw = AlertItem(id: UUID(),
+                             title: Text("Draw!"),
+                             message: Text("It's a draw"),
+                             buttonTitle: Text("Try again"))
+}
+
+
 struct Move {
     let player: Player // Who made the move?
     let boardIndex: Int // Where on the board was it?
@@ -27,8 +51,7 @@ struct Move {
 
 // MARK: ViewModel
 
-
-class TTTViewModel: ObservableObject {
+final class TTTViewModel: ObservableObject {
     // Published properties that the view can observe
     @Published var moves: [Move?] = Array(repeating: nil, count: 9)
     @Published var isGameBoardDisabled: Bool = false
@@ -169,7 +192,7 @@ class TTTViewModel: ObservableObject {
 
 // MARK: ContentView
 
-struct ContentView: View {
+struct GameView: View {
     // Use the view model
     @StateObject private var viewModel = TTTViewModel()
     
@@ -223,7 +246,7 @@ struct ContentView: View {
 }
 
 #Preview {
-    ContentView()
+    GameView()
 }
 
 // Game squares
