+++ TikTakToe/TikTakToe/TikTakToeApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct TikTakToeApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            GameView()
         }
     }
 }
